Change transaction amount to float

+++ src/Responses/Transaction.php
@@ -10,7 +10,7 @@ class Transaction
     public function __construct(
         public readonly string $email,
         public readonly string $bankId,
-        public readonly int $amount,
+        public readonly float $amount,
         public readonly bool $switching,
         public readonly string $orderCode,
         public readonly TransactionStatus $statusId,
